+++ common/dao-api/src/main/java/org/thingsboard/server/dao/mes/vo/TaskListFinishVo.java
@@ -46,7 +46,7 @@ public class TaskListFinishVo {
     private String bodyMaterialName;
 
     @ApiModelProperty("预期产量")
-    private Float billPlanQty;
+    private BigDecimal billPlanQty;
 
     @ApiModelProperty("明细-单位-编码")
     private String bodyUnit;
